* DXProject/DXProject.vcproj * DXProject/gocVisualSphere.h * DXProject/D3DDevice.cpp * DXProject/D3DDevice.h * DXProject/IVideoDevice.h * DXProject/MeshBuffer.h * DXProject/NullDriver.h * DXProject/vertex.h * DXProject/camera.cpp * DXProject/D3DSample.cpp * DXProject/math.h * DXProject/quaternion.h * DXProject/vector3d.h

+++ ProjectSin/DXProject/gocVisualSphere.h
@@ -18,7 +18,7 @@ public:
 	void setSize(const float r) { mSize = r; }
 
 private:
-	int mSize;
+	float mSize;
 	static goc_id_type mFamilyId;
 
 	ID3D10Buffer* vertexBuffer;
